+++ gradle.properties
@@ -6,11 +6,11 @@ minecraft_version=1.21.1
 yarn_mappings=1.21.1+build.3
 loader_version=0.18.0
 # Mod Properties
-mod_version=0.3.0-Beta
+mod_version=0.4.0-Beta
 maven_group=net.hana
 archives_base_name=cobblemon-paleontologist
 # Dependencies
 # check this on https://modmuss50.me/fabric.html
 fabric_version=0.116.7+1.21.1
 fabric_kotlin_version=1.13.4+kotlin.2.2.0
-cobblemon_version=1.7.1+1.21.1
+cobblemon_version=1.7.2+1.21.1
